Move all definitions into aggiemap common lib
Next step is to add ability to select different preset conneciton endpionts for dev/prod, exclusion for layers (veoride, 3d buildings), and popup overrides
