--- apps/aggiemap-angular/src/environments/definitions.ts
@@ -1,590 +1,0 @@
-import { SearchSource, SearchSourceQueryParamsProperties } from '@tamu-gisc/ui-kits/ngx/search';
-import { LayerSource, LegendItem } from '@tamu-gisc/common/types';
-import { Popups } from '@tamu-gisc/aggiemap/ngx/popups';
-
-import esri = __esri;
-
-export const Connections = {
-  basemapUrl: 'https://gis.tamu.edu/arcgis/rest/services/FCOR/TAMU_BaseMap/MapServer',
-  inforUrl: 'https://gis.tamu.edu/arcgis/rest/services/FCOR/MapInfo_20190529/MapServer',
-  accessibleUrl: 'https://gis.tamu.edu/arcgis/rest/services/FCOR/ADA_120717/MapServer/0',
-  constructionUrl: 'https://gis.tamu.edu/arcgis/rest/services/FCOR/Construction_2018/MapServer',
-  departmentUrl: 'https://gis.tamu.edu/arcgis/rest/services/FCOR/DepartmentSearch/MapServer/1',
-  tsMainUrl: 'https://gis.tamu.edu/arcgis/rest/services/TS/TS_Main/MapServer',
-  bikeRacksUrl: 'https://gis.tamu.edu/arcgis/rest/services/TS/TS_Bicycles/MapServer/3',
-  bikeLocationsUrl: 'https://veoride.geoservices.tamu.edu/api/vehicles/basic/geojson'
-};
-
-export const Definitions = {
-  BUILDINGS: {
-    id: 'buildings',
-    layerId: 'buildings-layer',
-    name: 'Buildings',
-    url: `${Connections.basemapUrl}/1`,
-    popupComponent: Popups.BuildingPopupComponent
-  },
-  CONSTRUCTION: {
-    id: 'construction_zone',
-    layerId: 'construction_zone-layer',
-    name: 'Construction Zone',
-    url: `${Connections.constructionUrl}`,
-    popupComponent: Popups.ConstructionPopupComponent
-  },
-  POINTS_OF_INTEREST: {
-    id: 'poi',
-    layerId: 'poi-layer',
-    name: 'Points of Interest',
-    url: `https://services1.arcgis.com/qr14biwnHA6Vis6l/ArcGIS/rest/services/AgMap_Statues/FeatureServer/0`,
-    popupComponent: Popups.PoiPopupComponent
-  },
-  RESTROOMS: {
-    id: 'restrooms',
-    layerId: 'restrooms-layer',
-    name: 'Restrooms',
-    url: `${Connections.inforUrl}/1`,
-    popupComponent: Popups.RestroomPopupComponent
-  },
-  LACTATION_ROOMS: {
-    id: 'lactation-rooms',
-    layerId: 'lactation-rooms-layer',
-    name: 'Lactation Rooms',
-    url: `${Connections.inforUrl}/2`,
-    popupComponent: Popups.LactationPopupComponent
-  },
-  SURFACE_LOTS: {
-    id: 'surface-lots',
-    layerId: 'surface-lots-layer',
-    name: 'Surface Lots',
-    url: `${Connections.basemapUrl}/9`,
-    popupComponent: Popups.ParkingLotPopupComponent
-  },
-  VISITOR_PARKING: {
-    id: 'visitor-parking',
-    layerId: 'visitor-parking-layer',
-    name: 'Visitor Parking',
-    url: `${Connections.inforUrl}/3`,
-    popupComponent: Popups.ParkingKioskPopupComponent
-  },
-  TRANSPORTATION_PARKING: {
-    id: 'transportation-parking',
-    layerId: 'transportation-parking-layer',
-    name: 'Transportation Parking',
-    url: `${Connections.tsMainUrl}/6`,
-    popupComponent: Popups.ParkingKioskPopupComponent
-  },
-  ACESSIBLE_ENTRANCES: {
-    id: 'accessible-entrances',
-    layerId: 'accessible-entrances-layer',
-    name: 'Accessible Entrances',
-    url: `${Connections.accessibleUrl}`,
-    popupComponent: Popups.AccessiblePopupComponent
-  },
-  EMERGENCY_PHONES: {
-    id: 'emergency-phones',
-    layerId: 'emergency-phones-layer',
-    name: 'Emergency Phones',
-    url: `${Connections.inforUrl}/4`
-  },
-  BIKE_RACKS: {
-    id: 'bike-racks',
-    layerId: 'bike-racks-layer',
-    name: 'Bike Racks',
-    url: `${Connections.bikeRacksUrl}`
-  },
-  BIKE_LOCATIONS: {
-    id: 'bike-locations',
-    layerId: 'bike-locations-layer',
-    name: 'VeoRide Bikes',
-    url: `${Connections.bikeLocationsUrl}`
-  }
-};
-
-export const commonLayerProps = {
-  outFields: ['*'],
-  minScale: 100000,
-  maxScale: 0,
-  elevationInfo: { mode: 'relative-to-ground', offset: 1 } as esri.FeatureLayerElevationInfo,
-  popupEnabled: false
-};
-
-// Persistent layer definitions that will be processed by a factory and added to the map.
-export const LayerSources: LayerSource[] = [
-  {
-    type: 'feature',
-    id: Definitions.BUILDINGS.layerId,
-    title: Definitions.BUILDINGS.name,
-    url: Definitions.BUILDINGS.url,
-    popupComponent: Definitions.BUILDINGS.popupComponent,
-    listMode: 'hide',
-    visible: true,
-    layerIndex: 1,
-    native: {
-      ...commonLayerProps,
-      legendEnabled: false,
-      renderer: {
-        type: 'simple',
-        symbol: {
-          type: 'simple-fill',
-          style: 'solid',
-          color: [0, 0, 0, 0.01],
-          outline: {
-            width: '0'
-          }
-        }
-      }
-    }
-  },
-  {
-    type: 'feature',
-    id: Definitions.CONSTRUCTION.layerId,
-    title: Definitions.CONSTRUCTION.name,
-    url: Definitions.CONSTRUCTION.url,
-    popupComponent: Definitions.CONSTRUCTION.popupComponent,
-    listMode: 'show',
-    visible: true,
-    layerIndex: 2,
-    native: {
-      ...commonLayerProps
-    }
-  },
-  {
-    type: 'feature',
-    id: Definitions.POINTS_OF_INTEREST.layerId,
-    title: Definitions.POINTS_OF_INTEREST.name,
-    url: Definitions.POINTS_OF_INTEREST.url,
-    popupComponent: Definitions.POINTS_OF_INTEREST.popupComponent,
-    listMode: 'show',
-    visible: false,
-    native: {
-      ...commonLayerProps,
-      renderer: {
-        type: 'simple',
-        symbol: {
-          type: 'picture-marker',
-          url: '/assets/images/markers/statue-icon.png',
-          width: '20px',
-          height: '30.2px'
-        }
-      }
-    }
-  },
-  {
-    type: 'feature',
-    id: Definitions.RESTROOMS.layerId,
-    title: Definitions.RESTROOMS.name,
-    url: Definitions.RESTROOMS.url,
-    popupComponent: Definitions.RESTROOMS.popupComponent,
-    listMode: 'show',
-    visible: false,
-    native: {
-      ...commonLayerProps
-    }
-  },
-  {
-    type: 'feature',
-    id: Definitions.LACTATION_ROOMS.layerId,
-    title: Definitions.LACTATION_ROOMS.name,
-    url: Definitions.LACTATION_ROOMS.url,
-    popupComponent: Definitions.LACTATION_ROOMS.popupComponent,
-    listMode: 'show',
-    visible: false,
-    native: {
-      ...commonLayerProps
-    }
-  },
-  {
-    type: 'feature',
-    id: Definitions.SURFACE_LOTS.layerId,
-    title: Definitions.SURFACE_LOTS.name,
-    url: Definitions.SURFACE_LOTS.url,
-    popupComponent: Definitions.SURFACE_LOTS.popupComponent,
-    listMode: 'hide',
-    visible: true,
-    layerIndex: 1,
-    native: {
-      ...commonLayerProps,
-      legendEnabled: false,
-      opacity: 0.0,
-      labelingInfo: [
-        {
-          symbol: {
-            type: 'text',
-            color: [0, 0, 0, 0]
-          }
-        }
-      ]
-    }
-  },
-  {
-    type: 'feature',
-    id: Definitions.VISITOR_PARKING.layerId,
-    title: Definitions.VISITOR_PARKING.name,
-    url: Definitions.VISITOR_PARKING.url,
-    popupComponent: Definitions.VISITOR_PARKING.popupComponent,
-    listMode: 'show',
-    visible: false,
-    native: {
-      ...commonLayerProps
-    }
-  },
-  {
-    type: 'feature',
-    id: Definitions.ACESSIBLE_ENTRANCES.layerId,
-    title: Definitions.ACESSIBLE_ENTRANCES.name,
-    url: Definitions.ACESSIBLE_ENTRANCES.url,
-    popupComponent: Definitions.ACESSIBLE_ENTRANCES.popupComponent,
-    listMode: 'show',
-    visible: false,
-    native: {
-      ...commonLayerProps
-    }
-  },
-  {
-    type: 'feature',
-    id: Definitions.EMERGENCY_PHONES.layerId,
-    title: Definitions.EMERGENCY_PHONES.name,
-    url: Definitions.EMERGENCY_PHONES.url,
-    listMode: 'show',
-    visible: false,
-    native: {
-      ...commonLayerProps
-    }
-  },
-  {
-    type: 'graphics',
-    id: 'selection-layer',
-    title: 'Selected Buildings',
-    listMode: 'hide',
-    visible: true,
-    popupComponent: Popups.BuildingPopupComponent,
-    native: {
-      ...commonLayerProps
-    }
-  },
-  {
-    type: 'graphics',
-    id: 'bus-route-layer',
-    title: 'Bus Routes',
-    listMode: 'hide',
-    visible: true,
-    native: {
-      ...commonLayerProps
-    }
-  }
-];
-
-// Static legend items
-export const LegendSources: LegendItem[] = [
-  {
-    id: 'university-building-legend',
-    title: 'University Building',
-    src: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAJBJREFUOI1jYaAyYIExAtnZGygxaP3Pnw1wAwPZ2RuE7azruYVFyDLsw/2HDCkvXhirKCvPgruQW1iEQUJJmWwXMv386cPAzOzDQlgpaWDUwFEDRw0cZgZ+uP+QbEO+vn3DwIts4PqfPxsCL1xg+HThAgMDAwODsISEMamGimtooLoQVuIyMDAwMDwkw7VQPQDNmyRazmV6EgAAAABJRU5ErkJggg=='
-  },
-  {
-    id: 'non-university-building-legend',
-    title: 'Non-University Building',
-    src: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAHFJREFUOI1jYaAyYIExmpqaGigxqK6urgFuYFNTU4OCBGe9EC8bWYY9ff2doampiaGurq4B7kIhXjYGZWk+sl348sMvBrgLqQlGDRw1cNTAYWbg09ffyTbk3edfqAbW1dU1NDU1wYsgcgBKAYssQCkAABa3ILBI1xApAAAAAElFTkSuQmCC'
-  },
-  {
-    id: 'off-campus-building-legend',
-    title: 'Off Campus Building',
-    src: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAGdJREFUOI3t1KENADEMA8CADhBc2H0802cnLxVUdYRHH7U0LXrVKAo4GbnI4ZTvIGk7EAALkKTVWh9VTWHuLiQFgEVDVZXWWrph712i4clc8IIX/Bno7mlkjLGCAIxkTFAmy8DOj9281XsgH/GbKEkAAAAASUVORK5CYII='
-  },
-  {
-    id: 'garage-parking-legend',
-    title: 'Garage Parking',
-    src: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAGtJREFUOI3t1LEOgCAMBNAOTMBgogNfezb3v4YQxdnJRldkMtzUdHi56Zx0jrsPkvoFAqAGktRlntYYfBO25SIkBYBawxi8pLQ0N9z3KtawZwY4wAH+DNxyaUaOer5BAErSJqglr4F9Pr7mAt2qILaQAOa8AAAAAElFTkSuQmCC'
-  },
-  {
-    id: 'restricted-legend',
-    title: 'Restricted 6am to 6pm',
-    src: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAFBJREFUOI1jYaAyYIExzq5Mb6DEIOPwmQ1wA8+uTG9gZPhfT4mBZ1emMxiHz2xgIayUNDBq4KiBowaOGkhHA43DZzacXZlOkUEoBSyyAKUAAGgHEnMWcHNYAAAAAElFTkSuQmCC'
-  },
-  {
-    id: 'paved-surface-legend',
-    title: 'Paved Surface',
-    src: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAGVJREFUOI3t1CEOACEMBMAKdF+A4aN7zf62AsMLTl0DtqAurGoqJqu2yOGU7yBpOxAAC5Ck1VofVU1h7i4kBYBFQ1WV1lq6Ye9douHJXPCCF/wZ6O5pZIyxggCMZExQJsvAzo/dvKA4IK9bweSrAAAAAElFTkSuQmCC'
-  },
-  {
-    id: 'sidewalk-legend',
-    title: 'Sidewalk',
-    src: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAEtJREFUOI1jYaAyYIExPr641UCJQfwSag1wA6GG1VNi4McXtxj4JdQaWAgrJQ2MGjhq4KiBowbS0UB+CbWGjy9uUWQQSgGLLEApAAAvoxCdulpoFwAAAABJRU5ErkJggg=='
-  },
-  {
-    id: 'railroad-legend',
-    title: 'Railroad',
-    src: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAHhJREFUOI3tkrENgDAMBL+I9EUmYA7WYCfCTLBG5qBz/x0VRUAkAtIg8q3t+7dlh8pyDfgnIMnwBiQpJEBJI8mpMDQAWEnGYkIA0Xs/m9llM8kkTQlYRQnQzEJuHUk9gG5PeqiFBEhyKbnnzE4Jc3e5ow89dgM+1gZwZSYCkyusZQAAAABJRU5ErkJggg=='
-  }
-];
-
-const commonQueryParams: Partial<SearchSourceQueryParamsProperties> = {
-  f: 'json',
-  resultRecordCount: 5,
-  outFields: '*',
-  outSR: 4326,
-  returnGeometry: true,
-  spatialRel: 'esriSpatialRelIntersects'
-};
-
-// Search sources used for querying features.
-export const SearchSources: SearchSource[] = [
-  {
-    source: 'building',
-    name: 'Building',
-    url: `${Connections.basemapUrl}/1`,
-    queryParams: {
-      ...commonQueryParams,
-      where: {
-        keys: ['Number', 'BldgAbbr', 'BldgName'],
-        operators: ['LIKE', 'LIKE', 'LIKE'],
-        wildcards: ['includes', 'includes', 'includes'],
-        transformations: ['UPPER', 'UPPER', 'UPPER']
-      },
-      scoringWhere: {
-        keys: ['BldgName', 'BldgAbbr'],
-        operators: ['LIKE', 'LIKE'],
-        wildcards: ['startsWith', 'startsWith'],
-        transformations: ['UPPER', 'UPPER']
-      }
-    },
-    scoringKeys: ['attributes.BldgAbbr', 'attributes.Number', 'attributes.BldgName'],
-    featuresLocation: 'features',
-    displayTemplate: '{attributes.BldgName} ({attributes.Number})',
-    popupComponent: Popups.BuildingPopupComponent,
-    searchActive: true
-  },
-  {
-    source: 'building-exact',
-    name: 'Building',
-    url: `${Connections.basemapUrl}/1`,
-    queryParams: {
-      ...commonQueryParams,
-      where: {
-        keys: ['Number'],
-        operators: ['=']
-      }
-    },
-    featuresLocation: 'features',
-    displayTemplate: '{attributes.BldgName} ({attributes.Number})',
-    popupComponent: Popups.BuildingPopupComponent,
-    searchActive: false
-  },
-  {
-    source: 'university-departments',
-    name: 'University Departments',
-    url: `${Connections.departmentUrl}`,
-    queryParams: {
-      ...commonQueryParams,
-      where: {
-        keys: ['DeptName', 'CollegeName', 'DeptAbbre'],
-        operators: ['LIKE', 'LIKE', 'LIKE'],
-        wildcards: ['includes', 'includes', 'includes'],
-        transformations: ['UPPER', 'UPPER', 'UPPER']
-      }
-    },
-    featuresLocation: 'features',
-    displayTemplate: '{attributes.DeptName}',
-    popupComponent: 'BasePopupComponent',
-    searchActive: false,
-    altLookup: {
-      source: 'building-exact',
-      reference: {
-        keys: ['attributes.HOME1']
-      }
-    }
-  },
-  {
-    source: 'university-departments-exact',
-    name: 'University Departments',
-    url: `${Connections.departmentUrl}`,
-    queryParams: {
-      ...commonQueryParams,
-      resultRecordCount: 100,
-      where: {
-        keys: ['HOME1'],
-        operators: ['=']
-      }
-    },
-    featuresLocation: 'features',
-    displayTemplate: '{attributes.DeptName}',
-    searchActive: false
-  },
-  {
-    source: 'all-parking',
-    name: 'Parking',
-    url: `${Definitions.TRANSPORTATION_PARKING.url}`,
-    queryParams: {
-      ...commonQueryParams,
-      returnGeometry: true,
-      resultRecordCount: '*',
-      outFields: `GIS.TS.ParkingLots.FAC_CODE,
-      GIS.TS.ParkingLots.LotName,
-      GIS.TS.SpacePnt_Count.UB,
-      GIS.TS.SpacePnt_Count.Visitor_H_C
-      `,
-      where: {
-        keys: ['1'],
-        operators: ['=']
-      }
-    },
-    featuresLocation: 'features',
-    displayTemplate: '{attributes.LotName}',
-    searchActive: false
-  },
-  {
-    source: 'visitor-parking',
-    name: 'Visitor Parking',
-    url: `${Definitions.TRANSPORTATION_PARKING.url}`,
-    queryParams: {
-      ...commonQueryParams,
-      returnGeometry: true,
-      resultRecordCount: '*',
-      outFields: `GIS.TS.ParkingLots.FAC_CODE,
-      GIS.TS.ParkingLots.LotName,
-      GIS.TS.ParkingLots.LotType
-      `,
-      where: {
-        keys: ['GIS.TS.ParkingLots.LotType', 'GIS.TS.ParkingLots.LotType'],
-        operators: ['=', '=']
-      }
-    },
-    featuresLocation: 'features',
-    displayTemplate: '{attributes.LotName}',
-    searchActive: false
-  },
-  {
-    source: 'night-parking',
-    name: 'Night Parking',
-    url: `https://gis.tamu.edu/arcgis/rest/services/TS/AVPVisBSUBVenNWRetNSCMed/MapServer/6`,
-    queryParams: {
-      ...commonQueryParams,
-      returnGeometry: true,
-      resultRecordCount: '*',
-      outFields: `GIS.TS.ParkingLots.FAC_CODE,
-      GIS.TS.ParkingLots.LotName,
-      GIS.TS.Lot_Use.Night_Lot`,
-      where: {
-        keys: ['Night_Lot'],
-        operators: ['=']
-      }
-    },
-    featuresLocation: 'features',
-    displayTemplate: '{attributes.LotName}',
-    searchActive: false
-  },
-  {
-    source: 'one-parking',
-    name: 'Parking',
-    url: `${Definitions.TRANSPORTATION_PARKING.url}`,
-    queryParams: {
-      ...commonQueryParams,
-      returnGeometry: true,
-      resultRecordCount: '*',
-      outFields: `GIS.TS.ParkingLots.AggieMap,
-      GIS.TS.ParkingLots.FAC_CODE,
-      GIS.TS.ParkingLots.LotName,
-      GIS.TS.SpacePnt_Count.UB`
-    },
-    featuresLocation: 'features',
-    displayTemplate: '{attributes.LotName}',
-    searchActive: false
-  },
-  {
-    source: 'parking-garage',
-    name: 'Parking Garage',
-    url: `${Connections.basemapUrl}/0`,
-    queryParams: {
-      ...commonQueryParams,
-      where: {
-        keys: ['LotName', 'Name'],
-        operators: ['LIKE', 'LIKE'],
-        wildcards: ['includes', 'includes'],
-        transformations: ['UPPER', 'UPPER']
-      }
-    },
-    featuresLocation: 'features',
-    displayTemplate: '{attributes.LotName}',
-    popupComponent: Popups.BuildingPopupComponent,
-    searchActive: true
-  },
-  {
-    source: 'parking-lot',
-    name: 'Parking Lot',
-    url: `${Connections.basemapUrl}/9`,
-    queryParams: {
-      ...commonQueryParams,
-      where: {
-        keys: ['LotName'],
-        operators: ['LIKE'],
-        wildcards: ['includes'],
-        transformations: ['UPPER']
-      }
-    },
-    featuresLocation: 'features',
-    displayTemplate: '{attributes.LotName}',
-    popupComponent: Popups.ParkingLotPopupComponent,
-    searchActive: true
-  },
-  {
-    source: 'points-of-interest',
-    name: 'Points of Interest',
-    url: Definitions.POINTS_OF_INTEREST.url,
-    queryParams: {
-      ...commonQueryParams,
-      where: {
-        keys: ['name'],
-        operators: ['LIKE'],
-        wildcards: ['includes'],
-        transformations: ['UPPER']
-      }
-    },
-    featuresLocation: 'features',
-    displayTemplate: '{attributes.name}',
-    popupComponent: Popups.PoiPopupComponent,
-    searchActive: true
-  },
-  {
-    source: 'bike-racks',
-    name: 'Bike Racks',
-    url: `${Connections.bikeRacksUrl}`,
-    queryParams: {
-      ...commonQueryParams,
-      resultRecordCount: 9999,
-      where: {
-        keys: ['1'],
-        operators: ['=']
-      }
-    },
-    featuresLocation: 'features',
-    displayTemplate: '{attributes.Type}',
-    searchActive: false
-  }
-];
-
-export const SelectionSymbols = {
-  polygon: {
-    type: 'simple-fill',
-    style: 'solid',
-    color: [252, 227, 0, 0.55],
-    outline: {
-      color: [252, 227, 0, 0.8],
-      width: '2px'
-    }
-  },
-  point: {
-    type: 'simple-marker',
-    style: 'circle',
-    size: 8,
-    outline: {
-      width: 2
-    }
-  },
-  multipoint: {
-    type: 'simple-marker',
-    style: 'circle',
-    size: 8,
-    outline: {
-      width: 2
-    }
-  }
-};
