+++ apps/aggiemap-angular/src/environments/environment.ts
@@ -4,15 +4,21 @@
 // The list of which env maps to which file can be found in `.angular-cli.json`.
 
 import { LayerSource } from '@tamu-gisc/common/types';
-import { LayerSources as LS, Definitions, commonLayerProps } from './definitions';
+import { LayerSources as LS, Definitions, commonLayerProps } from '@tamu-gisc/aggiemap/ngx/common';
 
 export const environment = {
   production: false
 };
 
-export { SearchSources, Connections, Definitions, LegendSources, SelectionSymbols } from './definitions';
+export {
+  SearchSources,
+  Connections,
+  Definitions,
+  LegendSources,
+  SelectionSymbols,
+  Polygons
+} from '@tamu-gisc/aggiemap/ngx/common';
 export * from './notification-events';
-export * from './polygons';
 
 export const LayerSources = [
   ...LS,
