+++ libs/aggiemap/ngx/common/src/index.ts
@@ -1,1 +1,20 @@
-export * from './lib/aggiemap-ngx-common.module';
+// Web GIS resource endpoints
+export * from './lib/connections';
+
+// Dictionary of map layer definitions
+export * from './lib/definitions';
+
+// Layer definitions
+export * from './lib/layer-sources';
+
+// Static legend items
+export * from './lib/legend-sources';
+
+// Search sources for search components and/or lookups
+export * from './lib/search-sources';
+
+// Default map feature selection symbol styles
+export * from './lib/selection-symbols';
+
+// Default trip planner boundary
+export * from './lib/polygons';
