--- libs/aggiemap/ngx/common/src/lib/aggiemap-ngx-common.module.ts
@@ -1,7 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@NgModule({
-  imports: [CommonModule]
-})
-export class AggiemapNgxCommonModule {}
