+++ libs/aggiemap/ngx/common/src/lib/definitions.ts
@@ -1,0 +1,86 @@
+import { Connections } from './connections';
+import { Popups } from '@tamu-gisc/aggiemap/ngx/popups';
+
+export const Definitions = {
+  BUILDINGS: {
+    id: 'buildings',
+    layerId: 'buildings-layer',
+    name: 'Buildings',
+    url: `${Connections.basemapUrl}/1`,
+    popupComponent: Popups.BuildingPopupComponent
+  },
+  CONSTRUCTION: {
+    id: 'construction_zone',
+    layerId: 'construction_zone-layer',
+    name: 'Construction Zone',
+    url: `${Connections.constructionUrl}`,
+    popupComponent: Popups.ConstructionPopupComponent
+  },
+  POINTS_OF_INTEREST: {
+    id: 'poi',
+    layerId: 'poi-layer',
+    name: 'Points of Interest',
+    url: `https://services1.arcgis.com/qr14biwnHA6Vis6l/ArcGIS/rest/services/AgMap_Statues/FeatureServer/0`,
+    popupComponent: Popups.PoiPopupComponent
+  },
+  RESTROOMS: {
+    id: 'restrooms',
+    layerId: 'restrooms-layer',
+    name: 'Restrooms',
+    url: `${Connections.inforUrl}/1`,
+    popupComponent: Popups.RestroomPopupComponent
+  },
+  LACTATION_ROOMS: {
+    id: 'lactation-rooms',
+    layerId: 'lactation-rooms-layer',
+    name: 'Lactation Rooms',
+    url: `${Connections.inforUrl}/2`,
+    popupComponent: Popups.LactationPopupComponent
+  },
+  SURFACE_LOTS: {
+    id: 'surface-lots',
+    layerId: 'surface-lots-layer',
+    name: 'Surface Lots',
+    url: `${Connections.basemapUrl}/9`,
+    popupComponent: Popups.ParkingLotPopupComponent
+  },
+  VISITOR_PARKING: {
+    id: 'visitor-parking',
+    layerId: 'visitor-parking-layer',
+    name: 'Visitor Parking',
+    url: `${Connections.inforUrl}/3`,
+    popupComponent: Popups.ParkingKioskPopupComponent
+  },
+  TRANSPORTATION_PARKING: {
+    id: 'transportation-parking',
+    layerId: 'transportation-parking-layer',
+    name: 'Transportation Parking',
+    url: `${Connections.tsMainUrl}/6`,
+    popupComponent: Popups.ParkingKioskPopupComponent
+  },
+  ACESSIBLE_ENTRANCES: {
+    id: 'accessible-entrances',
+    layerId: 'accessible-entrances-layer',
+    name: 'Accessible Entrances',
+    url: `${Connections.accessibleUrl}`,
+    popupComponent: Popups.AccessiblePopupComponent
+  },
+  EMERGENCY_PHONES: {
+    id: 'emergency-phones',
+    layerId: 'emergency-phones-layer',
+    name: 'Emergency Phones',
+    url: `${Connections.inforUrl}/4`
+  },
+  BIKE_RACKS: {
+    id: 'bike-racks',
+    layerId: 'bike-racks-layer',
+    name: 'Bike Racks',
+    url: `${Connections.bikeRacksUrl}`
+  },
+  BIKE_LOCATIONS: {
+    id: 'bike-locations',
+    layerId: 'bike-locations-layer',
+    name: 'VeoRide Bikes',
+    url: `${Connections.bikeLocationsUrl}`
+  }
+};
