+++ libs/aggiemap/ngx/common/src/lib/selection-symbols.ts
@@ -1,0 +1,27 @@
+export const SelectionSymbols = {
+  polygon: {
+    type: 'simple-fill',
+    style: 'solid',
+    color: [252, 227, 0, 0.55],
+    outline: {
+      color: [252, 227, 0, 0.8],
+      width: '2px'
+    }
+  },
+  point: {
+    type: 'simple-marker',
+    style: 'circle',
+    size: 8,
+    outline: {
+      width: 2
+    }
+  },
+  multipoint: {
+    type: 'simple-marker',
+    style: 'circle',
+    size: 8,
+    outline: {
+      width: 2
+    }
+  }
+};
